cli: migrate error component to typescript
