+++ packages/gatsby-cli/src/reporter/loggers/ink/cli.js
@@ -7,7 +7,7 @@ import Spinner from "../ink/components/spinner"
 import { ProgressBar } from "../ink/components/progress-bar"
 
 import { Message } from "../ink/components/messages"
-import Error from "./components/error"
+import { Error } from "./components/error"
 import Develop from "../ink/components/develop"
 
 const showProgress = isTTY()
